chore: 패키지 이동
service > service.db
